requirements.txt
streamlit
Pillow
cryptography
fastapi
pytest
